requirements.txt
google-generativeai
pymupdf
flask
langchain
langchain-community
langchain-core
faiss-cpu
sentence-transformers
requests
pydantic
# langchain-google-genai
